// src/ledgers/ibc-go/codec/client.d.ts
export interface ClientUpdateProposal {
  /** the title of the update proposal */
  title: string
  /** the description of the proposal */
  description: string
  /** the client identifier for the client to be updated if the proposal passes */
  subjectClientId: string
  /**
   * the substitute client identifier for the client standing in for the subject
   * client
   */
  substituteClientId: string
}
/**
 * UpgradeProposal is a gov Content type for initiating an IBC breaking
 * upgrade.
 */
export interface UpgradeProposal {
  title: string
  description: string
  plan?: Plan
  /**
   * An UpgradedClientState must be provided to perform an IBC breaking upgrade.
   * This will make the chain commit to the correct upgraded (self) client state
   * before the upgrade occurs, so that connecting chains can verify that the
   * new upgraded client is valid by verifying a proof on the previous version
   * of the chain. This will allow IBC connections to persist smoothly across
   * planned chain upgrades
   */
  upgradedClientState?: Any
}
// ...
export declare const ClientUpdateProposal: {
  encode(message: ClientUpdateProposal, writer?: _m0.Writer): _m0.Writer
  decode(
    input: _m0.Reader | Uint8Array,
    length?: number | undefined
  ): ClientUpdateProposal
  fromJSON(object: any): ClientUpdateProposal
  toJSON(message: ClientUpdateProposal): unknown
  fromPartial(object: DeepPartial<ClientUpdateProposal>): ClientUpdateProposal
}
export declare const UpgradeProposal: {
  encode(message: UpgradeProposal, writer?: _m0.Writer): _m0.Writer
  decode(
    input: _m0.Reader | Uint8Array,
    length?: number | undefined
  ): UpgradeProposal
  fromJSON(object: any): UpgradeProposal
  toJSON(message: UpgradeProposal): unknown
  fromPartial(object: DeepPartial<UpgradeProposal>): UpgradeProposal
}
// ...
declare type Builtin =
  | Date
  | Function
  | Uint8Array
  | string
  | number
  | undefined
  | Long
export declare type DeepPartial<T> = T extends Builtin
  ? T
  : T extends Array<infer U>
  ? Array<DeepPartial<U>>
  : T extends ReadonlyArray<infer U>
  ? ReadonlyArray<DeepPartial<U>>
  : T extends {}
  ? {
      [K in keyof T]?: DeepPartial<T[K]>
    }
  : Partial<T>